+++ app/scripts/controllers/popupcontrollers/getphone.js
@@ -21,9 +21,25 @@ angular.module('servu')
       if(obj.provider == 'facebook'){
         socialService.fbLogin(obj).then(function (res){
         if(res.status == 200){
+          if(res.data.user.phone_confirmed){
           localStorage.setItem('userDetail', JSON.stringify(res));
           credentialService.authed = true;
           $state.go('user.joblist');
+          }
+          else if(!res.data.user.phone_confirmed){
+            $scope.closeThisDialog();
+            ngDialog.open({
+              template: 'views/dialogTemplates/pinPopup.html',
+              resolve: {
+                token: function () {
+                  return obj;
+                }
+              },
+              showClose:false,
+              overlay: false,
+              controller: 'getPinCtrl'
+            });
+          }
 
         }
         }, function (err) {
@@ -47,9 +63,25 @@ angular.module('servu')
       else if(obj.provider == 'google'){
         socialService.googleLogin(obj).then(function (res) {
           if(res.status == 200){
-            localStorage.setItem('userDetail', JSON.stringify(res));
-            credentialService.authed = true;
-            $state.go('user.joblist');
+            if(res.data.user.phone_confirmed){
+              localStorage.setItem('userDetail', JSON.stringify(res));
+              credentialService.authed = true;
+              $state.go('user.joblist');
+            }
+            else if(!res.data.user.phone_confirmed){
+              $scope.closeThisDialog();
+              ngDialog.open({
+                template: 'views/dialogTemplates/pinPopup.html',
+                resolve: {
+                  token: function () {
+                    return obj;
+                  }
+                },
+                showClose:false,
+                overlay: false,
+                controller: 'getPinCtrl'
+              });
+            }
 
           }
         }, function (err) {
@@ -76,11 +108,26 @@ angular.module('servu')
         obj.provider = token.provider;
         socialService.twitterLogin(obj).then(function (res) {
           console.log($scope.phone, 'phone');
-          if (res.status == 200) {
-            localStorage.setItem('userDetail', JSON.stringify(res));
-            credentialService.authed = true;
-            $scope.closeThisDialog();
-            $state.go('user.joblist');
+          if(res.status == 200){
+            if(res.data.user.phone_confirmed){
+              localStorage.setItem('userDetail', JSON.stringify(res));
+              credentialService.authed = true;
+              $state.go('user.joblist');
+            }
+            else if(!res.data.user.phone_confirmed){
+              $scope.closeThisDialog();
+              ngDialog.open({
+                template: 'views/dialogTemplates/pinPopup.html',
+                resolve: {
+                  token: function () {
+                    return obj;
+                  }
+                },
+                showClose:false,
+                overlay: false,
+                controller: 'getPinCtrl'
+              });
+            }
 
           }
         }, function (err) {
